backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.4.1
qdrant-client==1.8.0
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1
pytest==7.4.3
httpx==0.25.2